Add unified usage string, show HTIF help text
This renames the `help` method to `usage` and uses longer strings to
show (more succinctly) the help text. This uses a format of the help text
that matches both Rocket Chip's emulator and htif's format. HTIF usage
options are now shown along with this. Help text is now shown on stdout
instead of stderr (to align with HTIF and the Rocket Chip emulator).

This new method, `usage`, will no longer exit by default. Instead it just
prints the help text.

References to the old help text (which were expected to exit) are replaced
with a call to help followed by an error (or a lambda that does this in
the case of option_parser_t::help).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>